Make CalculatorForTotalGpa class control all the total gpa calculation to get it by that ways (4 scale - % scale - letter)
add required methods to calculate the total gap for 4 scale type.
add required methods to calculate the total gap for % scale type.
add required methods to calculate the total gap as a letter.
make separate methods to shows the logs of each process above to track it anytime we want.